requirements.txt
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
dnspython==2.1.0
ecdsa==0.19.0
exceptiongroup==1.2.1
fastapi==0.110.2
fastapi-csrf-protect==0.2.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
motor==2.5.1
mysqlclient==2.2.4
passlib==1.7.4
pyasn1==0.6.0
pydantic==1.10.15
PyJWT==2.3.0
pymongo==3.13.0
python-dateutil==2.9.0.post0
python-decouple==3.3
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.40
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0